Implementing Dynamic Related Media
Used the year and title of the currently viewed media item and passed
that to the related picture and video components. Based off of that
information it filters down to a list of relevant videos and pictures
and then trims that down to a max of 5 of each and displays those below
the currently selected media item.